requirements.txt
asgiref==3.7.2
astroid==3.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
flake8==7.0.0
h11==0.14.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
outcome==1.3.0.post0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tomlkit==0.12.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
wsproto==1.2.0
